Marta

Documentation for Marta.

Marta.CTImages.rescale! — Method

rescale!(x::AbstractArray;
    interval=nothing, calibration=nothing, window=nothing)

Rescale array x to the interval specified by interval.

If calibration is not nothing, then rescaling is done with reference to the values given by calibration. In other words, minimum and maximum are assumed to be the values specified by calibration.

See also: rescale

Marta.CTImages.rescale! — Method

rescale!(x::AbstractArray, slope::Number, intercept::Number)

In place linear rescaling of x.

See also: rescale

Marta.CTImages.rescale — Method

rescale(x::AbstractArray;
    interval=nothing, calibration=nothing, window=nothing)

Rescale array x to the interval specified by interval.

If calibration is not nothing, then rescaling is done with reference to the values given by calibration. In other words, minimum and maximum are assumed to be the values specified by calibration.

See also: rescale!

Marta.CTImages.rescale — Method

rescale(x::AbstractArray, slope::Number, intercept::Number; window)

Linear rescaling of x as x * slope + intercept.

See also: rescale!

Marta.CTImages.rotate — Method

rotate(mat::AbstractMatrix, α::Real; <keyword arguments>)

Rotate matrix mat about the center of angle α given in degrees. If rows and cols are not given, the rotated matrix has the same dimensions of the original matrix.

Arguments

mat: matrix to rotate.

α: angle in degrees.

rows=nothing: number of rows of the rotated matrix.

cols=nothing: number of columns of the rotated matrix.

interpolation: interpolation strategy. By default is BilinearInterpolation.

Marta.CTIO.load_image — Method

load_image(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Load image from file f.

Marta.CTIO.load_image — Method

load_image(io::IO[, T=Float32]; <keyword arguments>) where {T<:Real}

Load image from stream io.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

nϕ=0: number of projections (rows in the file for row major ordering).

nd=0: number of detectors (columns in the file for row major ordering).

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: write_image, read_ct_image, write_ct_image

Marta.CTIO.load_sinogram — Method

load_sinogram(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Load sinogram from file f.

Marta.CTIO.load_sinogram — Method

load_sinogram(io::IO[, ::Type{T}=Float32]; <keyword arguments>) where {T<:Real}

Load sinogram from stream io.

This function is specifically designed to load a sinogram. Do not expect consistent result with other type of images. Use more general function instead, refer to the See Also section.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

nϕ=0: number of projections (rows in the file for row major ordering).

nd=0: number of detectors (columns in the file for row major ordering).

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: write_sinogram, read_ct_image, write_ct_image

Marta.CTIO.load_tomogram — Method

load_tomogram(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Load reconstructed image from file f.

Marta.CTIO.load_tomogram — Method

load_tomogram(io::IO[, T=Float32]; <keyword arguments>) where {T<:Real}

Load reconstructed image from stream io.

This function is specifically designed to load a tomogram. Do not expect consistent result with other type of images. Use more general function instead, refer to the See Also section.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

nϕ=0: number of projections (rows in the file for row major ordering).

nd=0: number of detectors (columns in the file for row major ordering).

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: write_tomogram, read_ct_image, write_ct_image

Marta.CTIO.read_ct_image — Method

read_ct_image(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Read CT image from file f.

Marta.CTIO.read_ct_image — Method

read_ct_image(io::IO[, T=Float32]; kwargs...) where {T<:Real}

Read CT image from the stream io.

Keyword arguments

header: read rows and columns from header in the file. [default: false]

rows: the image rows.

cols: the image columns.

header_type: type for the two values of the header. [default: Int64]

row_major: assume the image is stored in row major ordering.

See also: write_ct_image

Marta.CTIO.write_ct_image — Method

write_ct_image(f::AbstractString, image::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write CT image to file f.

Marta.CTIO.write_ct_image — Method

write_ct_image(io::IO, image::AbstractMatrix{T}; kwargs...) where {T<:Real}

Write CT image to the stream io.

Keyword arguments

header: if true, write header information with rows and columns. [default: false]

header_type: type for the two values of the header. [default: Int64]

row_major: write image in row major ordering.

See also: read_ct_image

Marta.CTIO.write_image — Method

write_image(f::AbstractString, image::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write input image to file f.

Marta.CTIO.write_image — Method

write_image(io::IO, image::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write input image to stream io.

Arguments

io: the stream where the image should be written to.

image: the image data matrix.

header: specify if the image header should be written.

row_major: specify if the image should be stored in row major order.

header_type=Int64: the type of the header data.

See Also: read_image, write_ct_image, read_ct_image.

Marta.CTIO.write_sinogram — Method

write_sinogram(f::AbstractString, sinog::AbstractMatrix{T}) where {T<:Real}

Write sinogram to file f.

Marta.CTIO.write_sinogram — Method

write_sinogram(io::IO, sinog::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write sinogram sinog to stream io.

This function is specifically designed to deal with a sinogram image.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: load_sinogram, read_ct_image, write_ct_image.

Marta.CTIO.write_tomogram — Method

write_tomogram(f::AbstractString, tomog::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write reconstructed image to file f.

Marta.CTIO.write_tomogram — Method

write_tomogram(io::IO, tomog::AbstractMatrix{T}; <keyword arguments>) where {T<:Real})

Write reconstructed image to stream io.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

Marta.CTScan.AbstractCTScanner — Type

abstract type AbstractCTScanner{<:AbstractArray,<:AbstractGeometry} end

Abstract type for tests in this suite.

Marta.CTScan.FBPScanner — Method

FBPScanner(gst::AbstractCTScanner; name::Optional{String}, study_id::Optional{String})

Construct a FBPScanner object from another AbstractCTScanner object. This allows to reuse data.

Marta.CTScan.FBPScanner — Method

FBPScanner(
    geometry::AbstractGeometry,
    image::AbstractTestImage;
    name::Optional{String},
    study_id::Optional{String}
)

Construct a FBPScanner object.

Marta.CTScan.FBPScanner — Method

FBPScanner(
    image::AbstractTestImage;
    name::Optional{String},
    study_id::Optional{String}
)

Construct a FBPScanner object.

Marta.CTScan.FBPScanner — Method

FBPScanner(
    geometry::AbstractGeometry,
    data::AbstractCTData;
    name::Optional{String},
    study_id::Optional{String}
)

Construct a FBPScanner object.

Marta.AbstractAlgorithms.project_image — Method

project_image(gst::AbstractCTScanner, alg::Optional{A}; <keyword arguments>)
    where {A <: AbstractProjectionAlgorithm}

Compute sinogram for the test gst.

Keyword arguments depend on the algorithm employed, please see the relative documentation.

Marta.AbstractAlgorithms.reconstruct_image — Method

reconstruct_image(gst::AbstractCTScanner; <keyword arguments>)

Reconstruct CT image from test gst.

Keyword arguments depend on the specific algorithm used, please see the relative documentation.

Marta.CTIO.load_image — Method

load_image(f::AbstractString, gst::AbstractCTScanner)

Load image from file f.

Marta.CTIO.load_image — Method

load_image(io::IO, gst::AbstractCTScanner)

Load image from stream io.

Marta.CTIO.load_sinogram — Method

load_sinogram(f::AbstractString, gst::AbstractCTScanner)

Load sinogram from file f.

Marta.CTIO.load_sinogram — Method

load_sinogram(io::IO, gst::AbstractCTScanner)

Load sinogram from stream io.

Marta.CTIO.load_tomogram — Method

load_tomogram(f::AbstractString, gst::AbstractCTScanner)

Load tomogram from file f.

Marta.CTIO.load_tomogram — Method

load_tomogram(io::IO, gst::AbstractCTScanner)

Load tomogram from stream io.

Marta.CTIO.write_image — Method

write_image(f::AbstractString, gst::AbstractCTScanner)

Write input image to file f.

Marta.CTIO.write_image — Method

write_image(io::IO, gst::AbstractCTScanner)

Write input image to stream io.

Marta.CTIO.write_sinogram — Method

write_sinogram(f::AbstractString, gst::AbstractCTScanner)

Write sinogram to file f.

Marta.CTIO.write_sinogram — Method

write_sinogram(io::IO, gst::AbstractCTScanner)

Write sinogram to stream io.

Marta.CTIO.write_tomogram — Method

write_tomogram(f::AbstractString, gst::AbstractCTScanner)

Write tomogram to file f.

Marta.CTIO.write_tomogram — Method

write_tomogram(io::IO, gst::AbstractCTScanner)

Write reconstructed image to stream io.

Marta.CTScan.compute_gray_scale — Method

compute_gray_scale(image::AbstractMatrix, imp::ImageParams; mean = false)

Compute the corresponding gray scale for the test image image.

Marta.CTScan.compute_gray_scale — Method

compute_gray_scale(gst::AbstractCTScanner, imp::ImageParams; mean = false)

Compute the corresponding gray scale for the reconstructed image.

Marta.CTScan.create_image! — Method

create_image!(gst::AbstractCTScanner, par::ImageParams)

Create gray scale image for the test gst.

Marta.CTScan.sample_sinogram — Method

sample_sinogram(gst::AbstractCTScanner; <keyword arguments>)

Returns a new test object with the resampled sinogram.

Marta.CalibrationBase.calibrate_image — Method

calibrate_image(gst::AbstractCTScanner; min_pos, max_pos, interval=0..1, window=nothing)

Perform calibration of input image with reference values.

Marta.CalibrationBase.calibrate_image — Method

calibrate_image(gst::AbstractCTScanner, imp::ImageParams; interval=nothing, window=nothing)

Perform calibration of input image using image parameters as reference.

Marta.CalibrationBase.calibrate_tomogram — Method

calibrate_tomogram(gst::AbstractCTScanner; min_pos, max_pos, interval=0..1, window=nothing)

Perform calibration of reconstructed image with reference values.

Marta.CalibrationBase.calibrate_tomogram — Method

calibrate_tomogram(gst::AbstractCTScanner, imp::ImageParams; interval=nothing, window=nothing)

Perform calibration of reconstructed image using image parameters as reference.

Marta.CTPlots.plot_gray_scale — Method

plot_gray_scale(
    imp::ImageParams,
    gray_scale_data...;
    options = nothing,
    <keyword arguments>
)

Plot gray scale data.

Additional keyword arguments are passed to the Plots library.

Marta.CTPlots.plot_image — Method

plot_image(gst::AbstractCTScanner)

Plot input image created for the test gst.

Marta.CTPlots.plot_image — Method

plot_image(image::AbstractMatrix{T}; <keyword arguments>) where {T}

Plot CT image image with predefined options.

Additional keyword arguments are passed to the Plots library functions.

See also: plot_sinogram, plot_tomogram

Marta.CTPlots.plot_sinogram — Method

plot_sinogram(gst::AbstractCTScanner)

Plot the sinogram computed for the test gst.

Marta.CTPlots.plot_sinogram — Method

plot_sinogram(sinog::AbstractMatrix{T}; <keyword arguments>) where {T}

Plot sinogram sinog with predefined options.

Additional keyword arguments are passed to the underlying plotting driver.

See also: plot_image, plot_tomogram

Marta.CTPlots.plot_tomogram — Method

plot_tomogram(gst::AbstractCTScanner)

Plot the reconstructed image for the test gst.

Marta.CTPlots.plot_tomogram — Method

plot_tomogram(tomog::AbstractMatrix{T}; <keyword arguments>) where {T}

Plot reconstructed CT image tomog with predefined options.

Alias for plot_image. See plot_image for full reference. Additional keyword arguments are passed to the Plots library.

See also: plot_sinogram, plot_image