Algorithms documentation

MartaCT.AbstractAlgorithms.iradon — Function

iradon(sinog::AbstractMatrix, xs[, ys[, algorithm::AbstractProjectionAlgorithm[, coo::AbstractCoordinates]]]; <keyword arguments>)

Compute the inverse Radon transform of sinog on the points given by the vectors xs and ys. If ys is omitted, the reconstruction is performed on the square with xs == ys. The default reconstruction algorithm is FBP. Please refer to the respective documentation for additional parameters.

See Also: reconstruct_image

MartaCT.AbstractAlgorithms.iradon — Method

iradon(sinog::AbstractMatrix[, geometry::AbstractGeometry[, params::AbstractParams[, algorithm::AbstractProjectionAlgorithm]]]; <keyword arguments>)

Compute the inverse Radon transform of sinog with explicit geometry. If the algorithm needs specific parameters, these can be passed with params. Additional parameters can be passed to the algorithm through keyword arguments. Please see the respective documentation for more details. The default algorithm is FBP.

See Also: reconstruct_image

MartaCT.AbstractAlgorithms.iradon — Method

iradon(sinog::AbstractMatrix[, algorithm::AbstractProjectionAlgorithm]; <keyword arguments>)

Compute the inverse Radon transform of sinog with parameters given as keyword arguments. The parameters depend on the algorithm, please see the relative documentation. The default algorithm is FBP.

See Also: reconstruct_image

MartaCT.AbstractAlgorithms.radon — Function

radon(image::AbstractMatrix, geometry::AbstractGeometry[, algorithm::AbstractProjectionAlgorithm]; <keyword arguments>)

Compute the Radon transform of image with explicit geometry. Additional parameters can be passed to the algorithm through keyword arguments. Please see the relative documentation. The default algorithm is Radon.

See Also: project_image

MartaCT.AbstractAlgorithms.radon — Function

radon(image::AbstractMatrix[, algorithm::AbstractProjectionAlgorithm]; <keyword arguments>)

Compute the Radon transform of image with parameters given as keyword arguments. The parameters depend on the algorithm, please see the relative documentation. The default algorithm is Radon.

See Also: project_image

MartaCT.AbstractAlgorithms.radon — Function

radon(image::AbstractMatrix, ts, ϕs[, algorithm::AbstractProjectionAlgorithm]; <keyword arguments>)

Compute the Radon transform of image with integration points given by the vector ts and projection angles given by the vector ϕs. Additional parameters can be passed to the algorithm through keyword arguments.Please see the relative documentation. The default algorithm is Radon.

See Also: project_image

MartaCT.RadonAlgorithm.radon_diag — Function

radon_diag(image::AbstractMatrix, ts::AbstractVector, ϕs::AbstractVector; <keyword arguments>)

Compute the Radon transform of image inside a circle of radius hypot(rows,cols)/2 where rows and cols are the dimensions of image.

See Also: radon_square

MartaCT.RadonAlgorithm.radon_square — Function

radon_square(image::AbstractMatrix, ts::AbstractVector, ϕs::AbstractVector; <keyword arguments>)

Compute the Radon transform of image inside the circle contained in the square of side min(rows,cols) where rows and cols are the dimensions of image.

See also: radon_default