Input/Output functions

MartaCT.CTIO.load_image — Method

load_image(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Load image from file f.

MartaCT.CTIO.load_image — Method

load_image(io::IO[, T=Float32]; <keyword arguments>) where {T<:Real}

Load image from stream io.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

nϕ=0: number of projections (rows in the file for row major ordering).

nd=0: number of detectors (columns in the file for row major ordering).

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: write_image, read_ct_image, write_ct_image

MartaCT.CTIO.load_sinogram — Method

load_sinogram(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Load sinogram from file f.

MartaCT.CTIO.load_sinogram — Method

load_sinogram(io::IO[, ::Type{T}=Float32]; <keyword arguments>) where {T<:Real}

Load sinogram from stream io.

This function is specifically designed to load a sinogram. Do not expect consistent result with other type of images. Use more general function instead, refer to the See Also section.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

nϕ=0: number of projections (rows in the file for row major ordering).

nd=0: number of detectors (columns in the file for row major ordering).

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: write_sinogram, read_ct_image, write_ct_image

MartaCT.CTIO.load_tomogram — Method

load_tomogram(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Load reconstructed image from file f.

MartaCT.CTIO.load_tomogram — Method

load_tomogram(io::IO[, T=Float32]; <keyword arguments>) where {T<:Real}

Load reconstructed image from stream io.

This function is specifically designed to load a tomogram. Do not expect consistent result with other type of images. Use more general function instead, refer to the See Also section.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

nϕ=0: number of projections (rows in the file for row major ordering).

nd=0: number of detectors (columns in the file for row major ordering).

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: write_tomogram, read_ct_image, write_ct_image

MartaCT.CTIO.read_ct_image — Method

read_ct_image(f::AbstractString[, T=Float32]; <keyword arguments>) where {T<:Real}

Read CT image from file f.

MartaCT.CTIO.read_ct_image — Method

read_ct_image(io::IO[, T=Float32]; kwargs...) where {T<:Real}

Read CT image from the stream io.

Keyword arguments

header: read rows and columns from header in the file. [default: false]

rows: the image rows.

cols: the image columns.

header_type: type for the two values of the header. [default: Int64]

row_major: assume the image is stored in row major ordering.

See also: write_ct_image

MartaCT.CTIO.write_ct_image — Method

write_ct_image(f::AbstractString, image::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write CT image to file f.

MartaCT.CTIO.write_ct_image — Method

write_ct_image(io::IO, image::AbstractMatrix{T}; kwargs...) where {T<:Real}

Write CT image to the stream io.

Keyword arguments

header: if true, write header information with rows and columns. [default: false]

header_type: type for the two values of the header. [default: Int64]

row_major: write image in row major ordering.

See also: read_ct_image

MartaCT.CTIO.write_image — Method

write_image(f::AbstractString, image::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write input image to file f.

MartaCT.CTIO.write_image — Method

write_image(io::IO, image::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write input image to stream io.

Arguments

io: the stream where the image should be written to.

image: the image data matrix.

header: specify if the image header should be written.

row_major: specify if the image should be stored in row major order.

header_type=Int64: the type of the header data.

See Also: read_image, write_ct_image, read_ct_image.

MartaCT.CTIO.write_sinogram — Method

write_sinogram(f::AbstractString, sinog::AbstractMatrix{T}) where {T<:Real}

Write sinogram to file f.

MartaCT.CTIO.write_sinogram — Method

write_sinogram(io::IO, sinog::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write sinogram sinog to stream io.

This function is specifically designed to deal with a sinogram image.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.

See Also: load_sinogram, read_ct_image, write_ct_image.

MartaCT.CTIO.write_tomogram — Method

write_tomogram(f::AbstractString, tomog::AbstractMatrix{T}; <keyword arguments>) where {T<:Real}

Write reconstructed image to file f.

MartaCT.CTIO.write_tomogram — Method

write_tomogram(io::IO, tomog::AbstractMatrix{T}; <keyword arguments>) where {T<:Real})

Write reconstructed image to stream io.

Arguments

io: the stream where to load the image from.

header=false: specify if the file contains the header.

row_major=false: tells if the data is stored in row major ordering or not.

header_type=Int64: the type of the header data.