CTImages

MartaCT.CTImages.rescale! — Method

rescale!(x::AbstractArray;
    interval=nothing, calibration=nothing, window=nothing)

Rescale array x to the interval specified by interval.

If calibration is not nothing, then rescaling is done with reference to the values given by calibration. In other words, minimum and maximum are assumed to be the values specified by calibration.

See also: rescale

MartaCT.CTImages.rescale! — Method

rescale!(x::AbstractArray, slope::Number, intercept::Number)

In place linear rescaling of x.

See also: rescale

MartaCT.CTImages.rescale — Method

rescale(x::AbstractArray;
    interval=nothing, calibration=nothing, window=nothing)

Rescale array x to the interval specified by interval.

If calibration is not nothing, then rescaling is done with reference to the values given by calibration. In other words, minimum and maximum are assumed to be the values specified by calibration.

See also: rescale!

MartaCT.CTImages.rescale — Method

rescale(x::AbstractArray, slope::Number, intercept::Number; window)

Linear rescaling of x as x * slope + intercept.

See also: rescale!

MartaCT.CTImages.rotate — Method

rotate(mat::AbstractMatrix, α::Real; <keyword arguments>)

Rotate matrix mat about the center of angle α given in degrees. If rows and cols are not given, the rotated matrix has the same dimensions of the original matrix.

Arguments

mat: matrix to rotate.

α: angle in degrees.

rows=nothing: number of rows of the rotated matrix.

cols=nothing: number of columns of the rotated matrix.

interpolation: interpolation strategy. By default is BilinearInterpolation.