CT Plots

MartaCT.CTPlots.plot_gray_scale — Method

plot_gray_scale(
    imp::ImageParams,
    gray_scale_data...;
    options = nothing,
    <keyword arguments>
)

Plot gray scale data.

Additional keyword arguments are passed to the Plots library.

MartaCT.CTPlots.plot_image — Method

plot_image(gst::AbstractCTScanner)

Plot input image created for the test gst.

MartaCT.CTPlots.plot_image — Method

plot_image(image::AbstractMatrix{T}; <keyword arguments>) where {T}

Plot CT image image with predefined options.

Additional keyword arguments are passed to the Plots library functions.

See also: plot_sinogram, plot_tomogram

MartaCT.CTPlots.plot_sinogram — Method

plot_sinogram(gst::AbstractCTScanner)

Plot the sinogram computed for the test gst.

MartaCT.CTPlots.plot_sinogram — Method

plot_sinogram(sinog::AbstractMatrix{T}; <keyword arguments>) where {T}

Plot sinogram sinog with predefined options.

Additional keyword arguments are passed to the underlying plotting driver.

See also: plot_image, plot_tomogram

MartaCT.CTPlots.plot_tomogram — Method

plot_tomogram(gst::AbstractCTScanner)

Plot the reconstructed image for the test gst.

MartaCT.CTPlots.plot_tomogram — Method

plot_tomogram(tomog::AbstractMatrix{T}; <keyword arguments>) where {T}

Plot reconstructed CT image tomog with predefined options.

Alias for plot_image. See plot_image for full reference. Additional keyword arguments are passed to the Plots library.

See also: plot_sinogram, plot_image