CT Scan abstraction

Marta.CTScan.AbstractCTScanner — Type

abstract type AbstractCTScanner{<:AbstractArray,<:AbstractGeometry} end

Abstract type for tests in this suite.

Marta.CTScan.FBPScanner — Method

FBPScanner(gst::AbstractCTScanner; name::Optional{String}, study_id::Optional{String})

Construct a FBPScanner object from another AbstractCTScanner object. This allows to reuse data.

Marta.CTScan.FBPScanner — Method

FBPScanner(
    geometry::AbstractGeometry,
    image::AbstractTestImage;
    name::Optional{String},
    study_id::Optional{String}
)

Construct a FBPScanner object.

Marta.CTScan.FBPScanner — Method

FBPScanner(
    image::AbstractTestImage;
    name::Optional{String},
    study_id::Optional{String}
)

Construct a FBPScanner object.

Marta.CTScan.FBPScanner — Method

FBPScanner(
    geometry::AbstractGeometry,
    data::AbstractCTData;
    name::Optional{String},
    study_id::Optional{String}
)

Construct a FBPScanner object.

Marta.AbstractAlgorithms.project_image — Method

project_image(gst::AbstractCTScanner, alg::Optional{A}; <keyword arguments>)
    where {A <: AbstractProjectionAlgorithm}

Compute sinogram for the test gst.

Keyword arguments depend on the algorithm employed, please see the relative documentation.

Marta.AbstractAlgorithms.reconstruct_image — Method

reconstruct_image(gst::AbstractCTScanner; <keyword arguments>)

Reconstruct CT image from test gst.

Keyword arguments depend on the specific algorithm used, please see the relative documentation.

Marta.CTIO.load_image — Method

load_image(f::AbstractString, gst::AbstractCTScanner)

Load image from file f.

Marta.CTIO.load_image — Method

load_image(io::IO, gst::AbstractCTScanner)

Load image from stream io.

Marta.CTIO.load_sinogram — Method

load_sinogram(f::AbstractString, gst::AbstractCTScanner)

Load sinogram from file f.

Marta.CTIO.load_sinogram — Method

load_sinogram(io::IO, gst::AbstractCTScanner)

Load sinogram from stream io.

Marta.CTIO.load_tomogram — Method

load_tomogram(f::AbstractString, gst::AbstractCTScanner)

Load tomogram from file f.

Marta.CTIO.load_tomogram — Method

load_tomogram(io::IO, gst::AbstractCTScanner)

Load tomogram from stream io.

Marta.CTIO.write_image — Method

write_image(f::AbstractString, gst::AbstractCTScanner)

Write input image to file f.

Marta.CTIO.write_image — Method

write_image(io::IO, gst::AbstractCTScanner)

Write input image to stream io.

Marta.CTIO.write_sinogram — Method

write_sinogram(f::AbstractString, gst::AbstractCTScanner)

Write sinogram to file f.

Marta.CTIO.write_sinogram — Method

write_sinogram(io::IO, gst::AbstractCTScanner)

Write sinogram to stream io.

Marta.CTIO.write_tomogram — Method

write_tomogram(f::AbstractString, gst::AbstractCTScanner)

Write tomogram to file f.

Marta.CTIO.write_tomogram — Method

write_tomogram(io::IO, gst::AbstractCTScanner)

Write reconstructed image to stream io.

Marta.CTScan.compute_gray_scale — Method

compute_gray_scale(image::AbstractMatrix, imp::ImageParams; mean = false)

Compute the corresponding gray scale for the test image image.

Marta.CTScan.compute_gray_scale — Method

compute_gray_scale(gst::AbstractCTScanner, imp::ImageParams; mean = false)

Compute the corresponding gray scale for the reconstructed image.

Marta.CTScan.create_image! — Method

create_image!(gst::AbstractCTScanner, par::ImageParams)

Create gray scale image for the test gst.

Marta.CTScan.sample_sinogram — Method

sample_sinogram(gst::AbstractCTScanner; <keyword arguments>)

Returns a new test object with the resampled sinogram.

Marta.CalibrationBase.calibrate_image — Method

calibrate_image(gst::AbstractCTScanner; min_pos, max_pos, interval=0..1, window=nothing)

Perform calibration of input image with reference values.

Marta.CalibrationBase.calibrate_image — Method

calibrate_image(gst::AbstractCTScanner, imp::ImageParams; interval=nothing, window=nothing)

Perform calibration of input image using image parameters as reference.

Marta.CalibrationBase.calibrate_tomogram — Method

calibrate_tomogram(gst::AbstractCTScanner; min_pos, max_pos, interval=0..1, window=nothing)

Perform calibration of reconstructed image with reference values.

Marta.CalibrationBase.calibrate_tomogram — Method

calibrate_tomogram(gst::AbstractCTScanner, imp::ImageParams; interval=nothing, window=nothing)

Perform calibration of reconstructed image using image parameters as reference.

Marta.CTScan.CTImageData.CTData — Type

mutable struct CTData{M}

Hold data for reconstruction. M should be a subtype of AbstractArray, for example Matrix{Float32} or AFMatrix{Float64}. By default, the library will use CPU objects rather than GPU ones.

Marta.CTScan.CTImageData.CTData — Method

CTData(::AbstractCTImage)

Construct a CTData object from its constituents.

Marta.CTScan.CTImageData.CTData — Method

CTData(::Type{M}) where {M<:AbstractArray{<:Number}}

Construct an empty CTData object.

Marta.CTScan.CTImageData.CTData — Method

CTData([::Type{T} = Float32]) where {T<:Number}

Construct an empty CTData object with eltype T.