Test images

MartaCT.TestImages.ImageParams — Type

struct ImageParams{T<:Real}

Hold information to construct the input test image.

MartaCT.TestImages.ImageParams — Method

ImageParams([T=Float32]; <keyword arguments>) where {T<:Real}

Construct ImageParams object.

Arguments:

width=200: width of the rectangle gray scale image.

height=40: height of the rectangle gray scale image.

pad=30: padding around image.

dist=10: distance between images.

radius=20: radius of calibration circle.

gray_scale=1000..1000: interval of gray scale values.

calibration_value=nothing: value of the calibration circle. If not specified, is gray_scale[2].

background=nothing: value to be used as background.

hounsfield=false: whether the gray scale should be in Hounsfield units.

MartaCT.TestImages.background_position — Function

background_position(imp::AbstractImageParams)

Return suitable position to calibrate background.

MartaCT.TestImages.calibration_position — Function

calibration_position(imp::AbstractImageParams)

Get the position for calibration (max value).

MartaCT.TestImages.circle_image — Function

circle_image(imp::AbstractImageParams)

Create a square image with a circle of given value from parameters imp.

MartaCT.TestImages.circle_image — Method

circle_image([T=Float32]; radius=20, value=1) where {T<:Real}

Create a square image with a circle of given value.

Examples

julia> circle_image(30, 0.8)
30×30 Array{Float32,2}:
[...]

See also: gray_scale_image, combine_images

MartaCT.TestImages.circle_polar_image — Method

circle_polar_image([T=Float32] nr, nϕ, radius; value=1) where {T<:Real}

Create a circle of radius radius inside a nr×nϕ image in polar coordinates .

Examples

julia> circle_polar_image(30, 30, 10)
30×30 Array{Float32,2}:
[...]

MartaCT.TestImages.circle_position — Function

circle_position(imp::AbstractImageParams)

Return circle position inside image given image parameters.

MartaCT.TestImages.combine_images — Method

combine_images(imp::ImageParams{T}, rect::AbstractMatrix{T}, circle::AbstractMatrix{T}) where {T <: Real}

Helper function to create a gray scale image.

Combine the rect image with the circle image for calibration. The circle image should be smaller than the gray scale image!

See also: gray_scale_image, circle_image

MartaCT.TestImages.create_image — Method

create_image(par::ImageParams)

Create gray scale image.

MartaCT.TestImages.create_pyramid_image — Method

create_pyramid_image(par::ImageParams)

Create pyramid gray scale image.

MartaCT.TestImages.gray_scale_image — Method

gray_scale_image(imp::ImageParams)

Create an image with a gray scale rectangle with given scale gray_scale from parameters imp.

MartaCT.TestImages.gray_scale_image — Method

gray_scale_image(
    [T=Float32];
    rows=40,
    cols=200,
    swidth=nothing,
    sheight=nothing,
    gray_scale=-1000..1000,
) where {T <: Real}

Create an image with a gray scale rectangle with given scale gray_scale.

Examples

julia> gray_scale_image(swidth=80, sheight=40, gray_scale=-1000..1000)
40×80 Array{Float32,2}:
[...]

See also: circle_image, combine_images

MartaCT.TestImages.gray_scale_indices — Method

gray_scale_indices(imp::ImageParams)

Return a tuple (row, column_range) where row is the row index of the scale and column_range is the range of columns.

MartaCT.TestImages.pyramid_gray_scale_image — Method

pyramid_gray_scale_image(imp::ImageParams)

Create an image with a pyramid gray scale rectangle with given scale gray_scale from parameters imp.

Create an image with a pyramid gray scale rectangle with given scale gray_scale from parameters imp.

MartaCT.TestImages.pyramid_gray_scale_image — Method

pyramidgrayscaleimage([T=Float32]; swidth=200, sheight=40, grayscale=-1000..1000, plateau = 0) where {T <: Real}

Create an image with a pyramid gray scale rectangle with given scale gray_scale.

Examples

julia> pyramid_gray_scale_image(swidth=80, sheight=40, gray_scale=-1000..1000)
40×80 Array{Float32,2}:
[...]

See also: gray_scale_image, circle_image, combine_images

MartaCT.TestImages.square_image — Method

square_image([T=Float32] r, c; l=nothing) where {T<:Real}

Create a l×l square inside a r×c image.

Examples

julia> square_image(30, 30; l=10)
30×30 Array{Float32,2}:
[...]

MartaCT.TestImages.square_position — Method

square_position(imp::SquareParams)

Return square position inside image given image parameters.